Fix compile-blocking type errors found by static audit
No Dart toolchain is available in this environment, so these were found by
auditing the patterns a compiler would reject rather than by building.

Type errors that would have failed the build:
- num.clamp returns num, not the receiver's type. Three call sites passed
  the result straight into a double or int parameter:
  thumbnail_capture (pixelRatio), restore_text_element (List.insert index)
  and invitation_page_renderer (Color.fromRGBO opacity).
- const AlwaysStoppedAnimation(1) infers AlwaysStoppedAnimation<int>, which
  FadeTransition rejects. Given an explicit <double>.

Version-fragile APIs replaced with portable equivalents:
- ref.mounted only exists in newer Riverpod 2.x. The voice notifier now
  tracks disposal itself, which matters because the speech engine fires
  callbacks after the dialog may have closed.
- A typed initialising formal (InvitationProject this.project) narrowed a
  nullable field; rewritten as a plain constructor parameter.
- The isolate payload from the style analyser is now a List<int> rather
  than a private class, so it does not depend on how a given Dart version
  handles sending arbitrary objects across isolates.

Also:
- The style-to-font mapping is now guarded by a debug assert that the
  suggested family exists in the catalogue; a typo there would silently
  render the fallback serif and make the suggestion a lie. This also puts
  the previously unused guard to work.
- The About page was an AppDialog, which always renders confirm *and*
  cancel — two buttons doing the same thing. Now a sheet with one Close.

Verification suite passes: 123 files, imports resolve, delimiters balanced,
design tokens exist, no unguarded num.clamp, no version-fragile APIs, 37
use cases and 13 repositories registered, routing complete, assets present,
no dead code.

+++ invitation_editor/lib/features/editor/domain/usecases/restore_text_element.dart
@@ -15,7 +15,9 @@ class RestoreTextElement {
     final elements = [...canvas.elements];
     // The list may have changed since the delete, so clamp rather than
     // trusting the old index blindly.
-    final insertAt = index.clamp(0, elements.length);
+    // toInt() matters: num.clamp returns num, which will not assign to the
+    // int that List.insert expects.
+    final insertAt = index.clamp(0, elements.length).toInt();
     elements.insert(insertAt, element);
 
     return canvas.copyWith(
